Move feed http handler to river package
